Split cluster upgrade tests into 9+3 node
For ease of running just one group.
Some of the refactoring should make this a bit nicer if we need to tweak it in future too.
