+++ cosmo_tester/test_suites/cluster/cfy_cluster_manager_resources/3_nodes_config.yaml
