+++ cosmo_tester/test_suites/cluster/cfy_cluster_manager_resources/9_nodes_config.yaml
